# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Backend API",
    "version": "1.0.0",
    "description": "Documentação da API do backend"
  },
  "servers": [
    {
      "url": "https://tematicoapi.zapchatbr.com/",
      "description": "API Temático"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "paths": {
...
    "/reservas": {
      "get": {
        "summary": "Listar reservas",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "codigoUH",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hoje",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de reservas"
          }
        },
        "tags": [
          "Reservas"
        ]
      },
      "post": {
        "summary": "Criar reserva",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idreservacm": {
                    "type": "integer"
                  },
                  "numeroreservacm": {
                    "type": "string"
                  },
                  "coduh": {
                    "type": "string"
                  },
                  "nome_hospede": {
                    "type": "string"
                  },
                  "contato": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "data_checkin": {
                    "type": "string",
                    "format": "date"
                  },
                  "data_checkout": {
                    "type": "string",
                    "format": "date"
                  },
                  "qtd_hospedes": {
                    "type": "integer"
                  }
                },
                "required": [
                  "coduh",
                  "nome_hospede",
                  "contato",
                  "email",
                  "data_checkin",
                  "data_checkout",
                  "qtd_hospedes"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reserva criada"
          }
        },
        "tags": [
          "Reservas"
        ]
      }
    },
    "/reservas/buscar": {
      "get": {
        "summary": "Buscar reserva por coduh e período",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "coduh",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "checkin",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "checkout",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reserva encontrada"
          },
          "404": {
            "description": "Reserva não encontrada"
          },
          "409": {
            "description": "Múltiplas reservas sobrepostas"
          }
        },
        "tags": [
          "Reservas"
        ]
      }
    },
    "/reservas/validate-coduh/{coduh}": {
      "get": {
        "summary": "Validar reserva ativa por coduh",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "coduh",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados da reserva e eventos relacionados"
          },
          "404": {
            "description": "Reserva não encontrada"
          }
        },
        "tags": [
          "Reservas"
        ]
      }
    },
    "/reservas/{id}/marcacoes": {
      "get": {
        "summary": "Listar marcações de uma reserva",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de marcações"
          },
          "404": {
            "description": "Reserva não encontrada"
          }
        },
        "tags": [
          "Reservas"
        ]
      }
    },
    "/reservas/{id}": {
      "put": {
        "summary": "Atualizar reserva",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idreservacm": {
                    "type": "integer"
                  },
                  "numeroreservacm": {
                    "type": "string"
                  },
                  "coduh": {
                    "type": "string"
                  },
                  "nome_hospede": {
                    "type": "string"
                  },
                  "contato": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "data_checkin": {
                    "type": "string",
                    "format": "date"
                  },
                  "data_checkout": {
                    "type": "string",
                    "format": "date"
                  },
                  "qtd_hospedes": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reserva atualizada"
          },
          "404": {
            "description": "Reserva não encontrada"
          }
        },
        "tags": [
          "Reservas"
        ]
      },
      "delete": {
        "summary": "Deletar reserva",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reserva deletada"
          },
          "404": {
            "description": "Reserva não encontrada"
          }
        },
        "tags": [
          "Reservas"
        ]
      }
    }
...
  }
}